impr: re-organize ryot_sprite animations
