--- crates/ryot_sprites/src/animations.rs
@@ -1,7 +1,0 @@
-use bevy_ecs::prelude::*;
-
-#[derive(Debug, Hash, PartialEq, Eq, Clone, SystemSet)]
-pub enum AnimationSystems {
-    Initialize,
-    Update,
-}
